ajout du deroule des exam

+++ database/migrations/2021_07_29_051945_create_exam_activities_table.php
@@ -15,7 +15,7 @@ class CreateExamActivitiesTable extends Migration
     {
         Schema::create('exam_activities', function (Blueprint $table) {
             $table->id();
-            $table->integer('duration');
+            $table->bigInteger('duration');
             $table->integer('order');
             $table->foreignId('exam_id')->constrained()->cascadeOnDelete();
             $table->boolean('archived')->default(false);
